Review improvements:
1. Repackaging to be consistent with "TCK Report Comparison"
2. Throw more informative exception messages if parallel execution is enabled
3. Mention `forkCount` in README
