+++ tools/tck-reporting/src/test/java/org/opencypher/tools/tck/reporting/cucumber/CucumberReportAdapterTest.java
@@ -25,7 +25,7 @@
  * described as "implementation extensions to Cypher" or as "proposed changes to
  * Cypher that are not yet approved by the openCypher community".
  */
-package org.opencypher.tools.tck.reporting;
+package org.opencypher.tools.tck.reporting.cucumber;
 
 import java.io.IOException;
 import java.io.InputStream;
